% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedeal.R
\name{bed2matrix}
\alias{bed2matrix}
\title{plink format(.bed, .bim, .fam) to matrix}
\usage{
bed2matrix(FilePrefix, ...)
}
\arguments{
\item{FilePrefix}{plink format(.bed, .bim, .fam) file's prefix}

\item{...}{NULL}
}
\value{
list
}
\description{
plink format(.bed, .bim, .fam) to matrix
}
