% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{circos.legendlink}
\alias{circos.legendlink}
\title{circos legend}
\usage{
circos.legendlink(
  x = 0.8,
  y = 0.5,
  legend.limits = c(min, max),
  legend.bin = 1,
  legend.col = legend.col,
  legend.cex = 0.5,
  legend.border = NA,
  legend.family = NULL,
  legend.lty = par("lty"),
  legend.lwd = par("lwd"),
  legend.font = 1,
  ...
)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{legend.limits}{legend's limits;c(min, max)}

\item{legend.bin}{int; legend's bin}

\item{legend.col}{A specification for the default plotting color}

\item{legend.cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default.}

\item{legend.border}{Color for the border of the legend}

\item{legend.family}{The name of a font family for drawing text}

\item{legend.lty}{line width for borders and shading}

\item{legend.lwd}{line width for borders and shading}

\item{legend.font}{An integer which specifies which font to use for text}

\item{...}{other graphical parameters}
}
\description{
circos legend
}
