% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{LDsave}
\alias{LDsave}
\title{save the plots with triangle heatmap}
\usage{
LDsave(
  filename,
  plots,
  width = NULL,
  height = NULL,
  dpi = 300,
  units = NULL,
  device = NULL,
  scale = 1,
  ...
)
}
\arguments{
\item{filename}{output filename of plot.}

\item{plots}{Plot to save.}

\item{width}{plot size.}

\item{height}{plot size.}

\item{dpi}{Plot resolution.}

\item{units}{"in", "cm", "mm", or "px".}

\item{device}{Device to use,Can either be a device function. one of "pdf",
"png", "jpg", "jepg", "tiff".}

\item{scale}{Multiplicative scaling factor.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\description{
save the plots with triangle heatmap
}
