% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recode}
\alias{recode}
\title{Convert character genotype to integer genotype}
\arguments{
\item{vector}{a character vector of genotypes}

\item{sep}{the separator for character genotypes(i.e., "" ,"/", "|")}
}
\description{
convert genotypes separated by ("", "/", "|", ...) to integer genotypes(0, 1, 2, NA, NA for missing value)
}
\examples{

g0 <- c('T/A', NA, 'T|T', NA, 'T|A', NA, 'T|T', 'T/A','T/T', 'T/T', 
        'T/A', 'A|A', 'T/T', 'T|A', 'T/A', 'T|T', NA, 'T|A', 'T/A', NA)
        
code012_g1 <- recode(g0, "/|")

g1 <- c('T/A', NA, 'T/T', NA, 'T/A', NA, 'T/T', 'T/A','T/T', 'T/T', 
        'T/A', 'A/A', 'T/T', 'T/A', 'T/A', 'T/T', NA, 'T/A', 'T/A', NA)  

code012_g1 <- recode(g1, "/")

g2 <- c('01', '01', '00', '01', '00', '01', '01', '01', '01', '00', 
        '01', '11', '01', '11', '01', '11','01', '01', '01', '11') 
        
code012_g2 <- recode(g2, "")
}
