% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedeal.R
\name{ped2matrix}
\alias{ped2matrix}
\alias{read.plink}
\alias{vcf2matrix}
\title{plink format(.ped, .map) to  matrix}
\usage{
ped2matrix(FilePrefix, ...)

read.plink(FilePrefix, ...)

vcf2matrix(file, ...)
}
\arguments{
\item{FilePrefix}{plink format(.bed, .bim, .fam) file's prefix}

\item{...}{NULL}

\item{file}{vcf filename}
}
\value{
list

list
}
\description{
plink format(.ped, .map) to  matrix

read plink format(.bed, .bim, .fam)

vcf to matrix
}
